+++ components/index.tsx
@@ -5,8 +5,6 @@ export { default as Headline } from './Headline/index';
 
 export { default as AvatarList } from './AvatarList';
 
-export { default as Charts } from './Charts';
-
 export { default as NumberInfo } from './NumberInfo';
 
 export { default as Trend } from './Trend';